Split georeferenced raster import boundaries
