+++ src/PlateauResoniteLink/Application/Importing/GeoTiffTagSnapshot.cs
@@ -1,0 +1,9 @@
+namespace PlateauResoniteLink.Application.Importing;
+
+internal sealed record GeoTiffTagSnapshot(
+    double[]? ModelTiePoint,
+    double[]? PixelScale,
+    double[]? ModelTransform,
+    ushort[]? GeoKeyDirectory,
+    double[]? GeoDoubleParams,
+    string? GeoAsciiParams);
